feat: Repository protocol Domain계층으로 분리
